==> 2023_Ath_multi-omics/Data_preprocessing/03_methylomic_matrix/2_single_site_methylation_P_A/08_get_overlapping_site_between_M_and_G.r <==
dat <- read.table('Methylation_sites_listgenome_wide_383_accessions.txt',head=F,sep='\t')
snp = read.table('All_SNPs_list.txt',head=F,sep='\t')
pos = regexpr('_C',dat[,1])
loc = substr(dat[,1],1,pos-1)
dat <- cbind(dat,loc)
shared <- t(t(intersect(loc,snp[,1])))
res <- merge(dat,shared,by.x='loc',by.y='V1')
write.table(unique(res$loc),'Overlapping_sites_between_M_and_G.txt',row.names=F,quote=F,sep='\t',col.names=F)
write.csv(res,'Methylation_genome_wide_383_accessions_overlapping_with_SNP.csv',row.names=F,quote=F)
